+++ device_calendar/lib/src/models/reminder.dart
@@ -1,9 +1,11 @@
 import 'package:flutter/foundation.dart';
 
 class Reminder {
+  /// The time when the reminder should be triggered expressed in terms of minutes before the start of the event
   int minutes;
 
-  Reminder({@required this.minutes});
+  Reminder({@required this.minutes})
+      : assert(minutes >= 0, 'Minutes must be greater than or equal than zero');
 
   Reminder.fromJson(Map<String, dynamic> json) {
     minutes = json['minutes'] as int;