+++ device_calendar/CHANGELOG.md
@@ -1,6 +1,9 @@
 # 1.0.0 TBD
 * Support for more advanced recurrence rules
-* **BREAKING CHANGE** `retrieveCalendars` and `retrieveEvents` now return lists that cannot be modified
+* Update README to include information about using ProGuard for issue [99](https://github.com/builttoroam/flutter_plugins/issues/99)
+* Made event title optional to fix issue [72](https://github.com/builttoroam/flutter_plugins/issues/72)
+* Return information about the organiser of the event as per issue [73](https://github.com/builttoroam/flutter_plugins/issues/73)
+* **BREAKING CHANGE** `retrieveCalendars` and `retrieveEvents` now return lists that cannot be modified as part of address issue [113](https://github.com/builttoroam/flutter_plugins/issues/113)
 
 # 0.2.2 19th August 2019
 * Add support for specifying the location of an event. Thanks to [oli06](https://github.com/oli06) and [zemanux](https://github.com/zemanux) for submitting PRs to add the functionality to iOS and Android respectively