+++ device_calendar/CHANGELOG.md
@@ -5,6 +5,7 @@
 * Return information about the organiser of the event as per issue [73](https://github.com/builttoroam/flutter_plugins/issues/73)
 * Return attendance status of attendees and if they're required for an event. These are details are different across iOS and Android and so are returned within platform-specific objects
 * Ability to modify attendees for an event
+* Ability to create reminders for events expressed in minutes before the event starts
 * **BREAKING CHANGE** `retrieveCalendars` and `retrieveEvents` now return lists that cannot be modified as part of address issue [113](https://github.com/builttoroam/flutter_plugins/issues/113)
 
 # 0.2.2 19th August 2019